--- services/app/src/components/views/FeaturedQuests.test.tsx
@@ -1,5 +1,0 @@
-describe('FeaturedQuests', () => {
-  test.skip('Does not show search in simulator', () => { /* TODO */ });
-
-  test.skip('Filters out Horror quests if Horror disabled', () => { /* TODO */ });
-});
