+++ services/app/src/components/views/Tutorials.test.tsx
@@ -1,0 +1,4 @@
+describe('Tutorials', () => {
+  test.skip('Filters out Horror quests if Horror disabled', () => { /* TODO */ });
+  test.skip('Filters out Future quests if Future disabled', () => { /* TODO */ });
+});
